Create fundamentals.py
This is one of the many ways to better understand what exactly is going on when you are writing code as a beginner. 
The better you understand the fundamentals the better you will understand the more advanced material in the future.
When we write a program consisting if multiple functions and multiple variable, it is imperative that wee understand the logic at work. 
This is one of the ways to accomplish that.